Moved pyramid3d series into a subfolder.
